Update yOCTScanAndFindTissueSurface.m
Fixes from https://github.com/WinetraubLab/myOCT/pull/28
PR 28, continued here according to feedback:

Resolved runtime issues caused by recent updates
Ensured compatibility with yOCTPhotobleachTile
No files created
Minor adjustments

+++ Applications/yOCTScanAndFindTissueSurface.m
@@ -6,6 +6,7 @@ function [surfacePosition_mm, x_mm, y_mm] = yOCTScanAndFindTissueSurface(varargi
 %   isVisualize - set to true to generate image heatmap visualization
 %       figure. Default is false
 %   octProbePath - Where is the probe.ini is saved to be used. Default 'probe.ini'.
+%   octProbeFOV_mm - How much of the field of view to use from the probe during scans.
 %   output_folder - Directory for temporary files, default './Surface_Analysis_Temp'. 
 %       These files will be deleted after analysis is completed.
 %   dispersionQuadraticTerm - Dispersion compensation parameter.
@@ -29,22 +30,24 @@ addParameter(p,'isVisualize',false);
 addParameter(p,'xRange_mm',[-1 1]);
 addParameter(p,'yRange_mm',[-1 1]);
 addParameter(p,'pixel_size_um',25);
+addParameter(p,'octProbeFOV_mm',[]);
 addParameter(p,'octProbePath','probe.ini',@ischar);
 addParameter(p,'output_folder','./Surface_Analysis_Temp');
 addParameter(p,'dispersionQuadraticTerm',79430000,@isnumeric);
 addParameter(p,'focusPositionInImageZpix',NaN,@isnumeric);
 addParameter(p,'v',false);
 addParameter(p,'skipHardware',false)
-addParameter(p, 'acceptableRange_mm', 0.025, @isnumeric)
 
 parse(p,varargin{:});
 in = p.Results;
+
 xRange_mm = in.xRange_mm;
 yRange_mm = in.yRange_mm;
 pixel_size_um = in.pixel_size_um;
 isVisualize = in.isVisualize;
+octProbeFOV_mm = in.octProbeFOV_mm;
 octProbePath = in.octProbePath;
-output_folder = in.output_folder;
+output_folder = [in.output_folder '\surfaceAnalysis_temp'];
 dispersionQuadraticTerm = in.dispersionQuadraticTerm;
 v = in.v;
 
@@ -59,6 +62,7 @@ yOCTScanTile (...
     volumeOutputFolder, ...
     xRange_mm, ...
     yRange_mm, ...
+    'octProbeFOV_mm', octProbeFOV_mm, ...
     'octProbePath', octProbePath, ...
     'pixelSize_um', pixel_size_um, ...
     'v',v,  ...
@@ -66,15 +70,15 @@ yOCTScanTile (...
     );
 
 if in.skipHardware
-    % No need to continue
-    surfacePosition_mm = 0;
-    x_mm = 0;
-    y_mm = 0;
-    return;
-end
+     % No need to continue
+     surfacePosition_mm = 0;
+     x_mm = 0;
+     y_mm = 0;
+     return;
+ end
 
 %% Check if focusPositionInImageZpix is provided, if not use yOCTFindFocusTilledScan
-if isempty(in.focusPositionInImageZpix)
+if isempty(in.focusPositionInImageZpix) || any(isnan(in.focusPositionInImageZpix))
     if (v)
         fprintf('%s Find focus position volume\n', datestr(datetime));
     end
@@ -105,19 +109,25 @@ end
 if (v)
     fprintf('%s Identifying tissue surface...\n', datestr(datetime));
 end
-tic;
+tSurface = tic;
 dimensions = yOCTChangeDimensionsStructureUnits(dimensions,'millimeters'); % Make sure it's in mm
-surfacePosition_mm = yOCTFindTissueSurface(logMeanAbs, dimensions, 'isVisualize', isVisualize);
+surfacePosition_mm = yOCTFindTissueSurface( ...
+    logMeanAbs, ...
+    dimensions, ...
+    'isVisualize', isVisualize, ...
+    'octProbeFOV_mm', octProbeFOV_mm);
+
 x_mm=dimensions.x.values;
 y_mm=dimensions.y.values;
-elapsedTimeSurfaceDetection_sec = toc;
+elapsedTimeSurfaceDetection_sec = toc(tSurface);
 if (v)
     fprintf('%s Surface identification completed in %.2f seconds.\n', datestr(datetime), elapsedTimeSurfaceDetection_sec);
 end
 
 %% Clean up
-if exist(output_folder, 'dir')
+if ~isempty(output_folder) && exist(output_folder, 'dir')
     rmdir(output_folder, 's'); % Remove the output directory after processing
+    output_folder = fileparts(output_folder); % Return to original output_folder
 end
 totalEndTime = datetime;  % Capture the ending time
 totalDuration = totalEndTime - totalStartTime;
